+++ src/info.rs
@@ -1,0 +1,23 @@
+use serde::{Deserialize, Serialize};
+use crate::{License, Contact};
+
+/// General information about the API.
+/// https://github.com/OAI/OpenAPI-Specification/blob/master/versions/2.0.md#info-object
+#[derive(Clone, Debug, Deserialize, Serialize, PartialEq, Default)]
+#[serde(rename_all = "lowercase")]
+pub struct Info {
+    /// A unique and precise title of the API.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub title: Option<String>,
+    /// A semantic version number of the API.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub description: Option<String>,
+    #[serde(rename = "termsOfService", skip_serializing_if = "Option::is_none")]
+    pub terms_of_service: Option<String>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub contact: Option<Contact>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub license: Option<License>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub version: Option<String>,
+}