+++ src/paths.rs
@@ -1,0 +1,26 @@
+use serde::{Deserialize, Serialize};
+use crate::{ReferenceOr, Operation, Parameter};
+use indexmap::IndexMap;
+
+/// https://github.com/OAI/OpenAPI-Specification/blob/master/versions/2.0.md#path-item-object
+#[derive(Clone, Debug, Deserialize, Serialize, PartialEq, Default)]
+pub struct PathItem {
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub get: Option<Operation>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub post: Option<Operation>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub put: Option<Operation>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub patch: Option<Operation>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub delete: Option<Operation>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub options: Option<Operation>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub head: Option<Operation>,
+    #[serde(default, skip_serializing_if = "Vec::is_empty")]
+    pub parameters: Vec<ReferenceOr<Parameter>>,
+}
+
+pub type Paths = IndexMap<String, ReferenceOr<PathItem>>;