+++ tests/common/mod.rs
@@ -1,4 +1,4 @@
-use autorust_openapi::v2::Spec;
+use autorust_openapi::OpenAPI;
 use std::{fs::File, io::Read};
 
 pub type Error = Box<dyn std::error::Error>;
@@ -11,7 +11,7 @@ pub fn assert_deserialize_without_ignored(paths: Vec<&str>) -> Result<()> {
         File::open(path)?.read_to_end(&mut bytes)?;
         let deserializer = &mut serde_json::Deserializer::from_slice(&bytes);
         let mut ignored = Vec::new();
-        let _spec: Spec = serde_ignored::deserialize(deserializer, |path| {
+        let _spec: OpenAPI = serde_ignored::deserialize(deserializer, |path| {
             let path = path.to_string();
             println!("    ignored {}", &path);
             ignored.push(path);